refactor : App 컴포넌트에서 orderData 불러오는 로직 훅으로 분리
- 내부 코드가 길어서 알아보기 힘듦
